+++ modules/server/src/network/resolvers/index.ts
@@ -1,13 +1,12 @@
-import { type GraphQLResolveInfo } from 'graphql';
+import { Resolver } from '@/gqlServer';
 import { isSuccess } from '../result';
-import { NodeConfig } from '../types/types';
+import { NodeConfig } from '../setup/query';
 import { resolveInfoToMap } from '../utils/gql';
 import { convertToSqon } from '../utils/sqon';
-import { aggregationPipeline } from './aggregations';
-import { NetworkNode } from './networkNode';
+import { aggregationPipeline, NetworkNode } from './aggregations';
 import { createResponse } from './response';
 
-export type NetworkSearchRoot = {
+type NetworkSearchRoot = {
 	nodes: NetworkNode[];
 	aggregations: Record<string, unknown>;
 };
@@ -15,14 +14,15 @@ export type NetworkSearchRoot = {
 /*
  * Type should match the "Network" GQL type definition arg types
  */
-export type NetworkAggregationArgs = {
+// TODO: shared?
+type NetworkAggregationArgs = {
 	filters?: object;
 	aggregations_filter_themselves?: boolean;
 	include_missing?: boolean;
 };
 
 /**
- * Create GQL resolvers.
+ * Resolvers for network search.
  *
  * It's important to have both remote connection data and aggregations under a single field
  * as remote connection data is dependant on aggregations query
@@ -32,38 +32,44 @@ export type NetworkAggregationArgs = {
  * @returns
  */
 export const createResolvers = (configs: NodeConfig[]) => {
+	const network: Resolver<NetworkSearchRoot, NetworkAggregationArgs, any> = async (
+		_unusedParentObj,
+		args,
+		_unusedContext,
+		info,
+	) => {
+		const requestedFieldsMap = resolveInfoToMap(info, 'aggregations');
+
+		/*
+		 * Checks validity of SQON
+		 * For now we will pass through the non SQON object to the pipeline
+		 *
+		 * TODO: resolve Arranger / SQONBuilder SQON outer wrapper conflict
+		 * {"content": [{...}], "op": "and"}
+		 */
+		if ('filters' in args) {
+			const result = convertToSqon(args.filters);
+			if (!isSuccess(result)) {
+				throw new Error(`${result.status} : ${result.message}`);
+			}
+		}
+		const queryVariables = { ...args };
+
+		/*
+		 * Aggregation pipeline entrypoint
+		 */
+		const { aggregationResults, nodeInfo } = await aggregationPipeline(
+			configs,
+			requestedFieldsMap,
+			queryVariables,
+		);
+
+		return createResponse({ aggregationResults, nodeInfo });
+	};
+
 	return {
 		Root: {
-			network: async (
-				parent: NetworkSearchRoot,
-				// as mentioned above, type should match gql typedefs
-				args: NetworkAggregationArgs,
-				context: unknown,
-				info: GraphQLResolveInfo,
-			) => {
-				const requestedFieldsMap = resolveInfoToMap(info, 'aggregations');
-
-				/*
-				 * checks validity of SQON
-				 * for now we will pass through the non SQON object to the pipeline
-				 * TODO: resolve Arranger / SQONBuilder SQON outer wrapper conflict
-				 * ie. {"content": [{...}], "op": "and"}
-				 */
-				if ('filters' in args) {
-					const result = convertToSqon(args.filters);
-					if (!isSuccess(result)) {
-						throw new Error(`${result.status} : ${result.message}`);
-					}
-				}
-				const queryVariables = { ...args };
-
-				const { aggregationResults, nodeInfo } = await aggregationPipeline(
-					configs,
-					requestedFieldsMap,
-					queryVariables,
-				);
-				return createResponse({ aggregationResults, nodeInfo });
-			},
+			network,
 		},
 	};
 };
